% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_interpolated_map.R
\name{download_interpolated_map}
\alias{download_interpolated_map}
\title{Download an interpolated raster map collected by a specific sensor on a selected acquisition date and variable}
\usage{
download_interpolated_map(id_sensor, date, variable, filename, token)
}
\arguments{
\item{id_sensor}{The id sensor that you want to analyze}

\item{date}{The date of the data acquired (string)}

\item{variable}{Can be count, height, width, area}

\item{filename}{Path + filename where the ZIP will be saved}

\item{token}{The Web JSON Token API (Bearer token)}
}
\value{
A list with:
\itemize{
\item status_code
\item zip_file (absolute path or NULL)
\item raw_response (text in case of error)
}
}
\description{
Download a ZIP file containing interpolated raster data for a specific sensor/date/variable.
}
