% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_photo.R
\name{download_photo}
\alias{download_photo}
\title{Download all the data and photo collected by a specific sensor on a selected acquisition date}
\usage{
download_photo(id_sensor, date, filename, token)
}
\arguments{
\item{id_sensor}{The id sensor that you want to analyze}

\item{date}{The date of the data acquired (string)}

\item{filename}{Path + filename where the ZIP will be saved}

\item{token}{The Web JSON Token API (Bearer token)}
}
\value{
Returns a ZIP file containing all the data and photo collected by a specific sensor on a selected acquisition date
}
\description{
Download a ZIP file containing all the data and photo collected by a specific sensor on a selected acquisition date
}
