% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_prescription_map.R
\name{download_prescription_map}
\alias{download_prescription_map}
\title{Download a ZIP file containing the prescription map in wgs84 related to a specific sensor on a selected acquisition date}
\usage{
download_prescription_map(
  id_sensor,
  date,
  automatic,
  zone,
  fertilizer,
  strategy,
  percentage,
  filename,
  token
)
}
\arguments{
\item{id_sensor}{The id sensor that you want to analyze}

\item{date}{The date of the data acquired (string)}

\item{automatic}{Must be set as True or False. If True means that number and position of zone management are setted automatically; If False means that number of zone management are setted by the user by definig the zone parameter}

\item{zone}{if AutoPrescription is setted False, here you can specify the number of zone}

\item{fertilizer}{set the amount of fertilizer to provide on avarage to the field}

\item{strategy}{Can be set as highwherehigh or highwherelow. If highwherehigh means provide an higher amout of fertilizer where the production level is higher. If highwherelow means provide an higher amout of fertilizer where the production levels is lower}

\item{percentage}{Set the difference percentange of the dose that you want to apply between the zones. you can set a value between 0 to 100. if you set 0 an automatic process will define the difference of the dose based on production levels.}

\item{filename}{Path + filename where the ZIP will be saved}

\item{token}{The Web JSON Token API (Bearer token)}
}
\value{
Download a ZIP file containing the prescription map in wgs84 related to a specific sensor on a selected acquisition date
}
\description{
Download a ZIP file containing the prescription map in wgs84 related to a specific sensor on a selected acquisition date
}
