% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_data.R
\name{get_all_data}
\alias{get_all_data}
\title{Get all the data related to all the sensors connected to your account}
\usage{
get_all_data(token)
}
\arguments{
\item{token}{The Web JSON Token API}
}
\value{
Returns a Dataframe with all the sensor data related to your account
}
\description{
Get all the data related to all the sensors connected to your account
}
