% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_acquisition.R
\name{get_date_acquisition}
\alias{get_date_acquisition}
\title{Get all the dates of the data acquisition related to one sensor connected to your account}
\usage{
get_date_acquisition(id_sensor, token)
}
\arguments{
\item{id_sensor}{The id sensor that you want to analyze}

\item{token}{The Web JSON Token API}
}
\value{
Returns a Dataframe with the sensor dates
}
\description{
Get all the dates of the data acquisition related to one sensor connected to your account
}
