% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_html_map.R
\name{get_html_map}
\alias{get_html_map}
\title{Retrieve the html widget leaflet map with the interpolated map related to a specific sensor on a selected acquisition date}
\usage{
get_html_map(
  id_sensor,
  date,
  variable,
  token,
  save_dir = ".",
  filename = NULL,
  open_in_browser = FALSE
)
}
\arguments{
\item{id_sensor}{The id sensor that you want to analyze}

\item{date}{The date of the data acquired (string)}

\item{variable}{Can be "count", "height", "width", "area"}

\item{token}{The Web JSON Token API (Bearer token)}

\item{save_dir}{The path string where to save the html widget (default = ".")}

\item{filename}{The file name string of the html file to be saved (optional, will be auto-generated if NULL)}

\item{open_in_browser}{Open in browser the html widget to be visualized. Default is FALSE}
}
\value{
Returns a list with status_code, html_file (or NULL), dataframe (or NULL) and raw_response.
}
\description{
Retrieve the html widget leaflet map with the interpolated map related to a specific sensor on a selected acquisition date
}
