% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_html_prescription_map.R
\name{get_html_prescription_map}
\alias{get_html_prescription_map}
\title{Retrieve the html widget leaflet map with the prescription map related to a specific sensor on a selected acquisition date}
\usage{
get_html_prescription_map(
  id_sensor,
  date,
  automatic,
  zone,
  fertilizer,
  strategy,
  percentage,
  save_dir,
  filename,
  token,
  open_in_browser = FALSE
)
}
\arguments{
\item{id_sensor}{The id sensor that you want to analyze}

\item{date}{The date of the data acquired (string)}

\item{automatic}{Must be set as True or False. If True means that number and position of zone management are setted automatically; If False means that number of zone management are setted by the user by definig the zone parameter}

\item{zone}{if AutoPrescription is setted False, here you can specify the number of zone}

\item{fertilizer}{set the amount of fertilizer to provide on avarage to the field}

\item{strategy}{Can be set as highwherehigh or highwherelow. If highwherehigh means provide an higher amout of fertilizer where the production level is higher. If highwherelow means provide an higher amout of fertilizer where the production levels is lower}

\item{percentage}{Set the difference percentange of the dose that you want to apply between the zones. you can set a value between 0 to 100. if you set 0 an automatic process will define the difference of the dose based on production levels.}

\item{save_dir}{The path string where to save the html widget (default = ".")}

\item{filename}{Filename where the html will be saved}

\item{token}{The Web JSON Token API (Bearer token)}

\item{open_in_browser}{Open in browser the html widget to be visualized. Default is FALSE}
}
\value{
Returns a list with status_code, html_file (or NULL), dataframe (or NULL) and raw_response.
}
\description{
Retrieve the html widget leaflet map with the prescription map related to a specific sensor on a selected acquisition date
}
