% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_data.R
\name{get_sensor_data}
\alias{get_sensor_data}
\title{Get all the data related to one sensor connected to your account}
\usage{
get_sensor_data(id_sensor, token)
}
\arguments{
\item{id_sensor}{The id sensor that you want to analyze}

\item{token}{The Web JSON Token API}
}
\value{
Returns a Dataframe with the data related to one sensor connected to your account
}
\description{
Get all the data related to one sensor connected to your account
}
