% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_data_by_date.R
\name{get_sensor_data_by_date}
\alias{get_sensor_data_by_date}
\title{Retrieve all data collected by a specific sensor on a selected acquisition date.}
\usage{
get_sensor_data_by_date(id_sensor, date, token)
}
\arguments{
\item{id_sensor}{The id sensor that you want to analyze}

\item{date}{The date of the data acquired}

\item{token}{The Web JSON Token API}
}
\value{
Returns a Dataframe with the sensor data related to a specific sensor on a selected acquisition date.
}
\description{
Retrieve all data collected by a specific sensor on a selected acquisition date.
}
