% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_list.R
\name{get_sensor_list}
\alias{get_sensor_list}
\title{Get the sensors list connected to your account}
\usage{
get_sensor_list(token)
}
\arguments{
\item{token}{The Web JSON Token API}
}
\value{
Returns a list of the sensors connected to your account
}
\description{
Get the sensors list connected to your account
}
