% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statistics.R
\name{get_statistics}
\alias{get_statistics}
\title{Get the descriptive statiscs related to one sensor connected to your account}
\usage{
get_statistics(id_sensor, variable, token)
}
\arguments{
\item{id_sensor}{The id sensor that you want to analyze}

\item{variable}{Can be "count", "height", "width", "area"}

\item{token}{The Web JSON Token API}
}
\value{
Returns a list of the statistics
}
\description{
Get the descriptive statiscs related to one sensor connected to your account
}
