% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get the API Token to interact with the FruitKount API Service}
\usage{
get_token(email, password)
}
\arguments{
\item{email}{Email of the user}

\item{password}{Password of the user}
}
\value{
The API Token String
}
\description{
This function allows you to get the API Token to interact with the FruitKount API Service
}
