% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_detail.R
\name{user_detail}
\alias{user_detail}
\title{Get the User Details}
\usage{
user_detail(token)
}
\arguments{
\item{token}{Email of the user}
}
\value{
The dataframe of the user detail
}
\description{
This function allows you to get the API User Details
}
